ZOOKEEPER-3978 : PR Comments Fixes
