+++ zookeeper-server/src/main/java/org/apache/zookeeper/server/ZooKeeperServerMXBean.java
@@ -180,11 +180,11 @@ public interface ZooKeeperServerMXBean {
      */
     void resetFsyncThresholdExceedCount();
     /**
-     * Reset resetNonMTLSRemoteConnCount statistics only.
+     * Reset NonMTLSRemoteConnCount statistics only.
      */
     void resetNonMTLSConnCount();
     /**
-     * Reset resetNonMTLSRemoteConnCount statistics only.
+     * Reset NonMTLSRemoteConnCount statistics only.
      */
     void resetAuthFailedCount();
     /**
