+++ zookeeper-server/src/test/java/org/apache/zookeeper/server/NettyServerCnxnTest.java
@@ -30,7 +30,11 @@ import static org.junit.jupiter.api.Assertions.fail;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import io.netty.channel.*;
+import io.netty.channel.Channel;
+import io.netty.channel.ChannelFuture;
+import io.netty.channel.ChannelHandlerContext;
+import io.netty.channel.ChannelId;
+import io.netty.channel.ChannelPipeline;
 import io.netty.util.Attribute;
 import io.netty.util.AttributeKey;
 import java.io.IOException;
@@ -195,7 +199,6 @@ public class NettyServerCnxnTest extends ClientBase {
         Mockito.doReturn(atr).when(channel).attr(
                 Mockito.any()
         );
-       // when(channel.attr(Mockito.any(AttributeKey.class))).thenReturn(atr);
         doNothing().when(atr).set(Mockito.any());
 
         ServerStats.Provider providerMock = mock(ServerStats.Provider.class);
