+++ zookeeper-server/src/test/java/org/apache/zookeeper/server/X509AuthFailureTest.java
@@ -17,7 +17,6 @@
  */
 package org.apache.zookeeper.server;
 
-//import com.salesforce.sds.keystore.DynamicKeyStoreProvider;
 import org.apache.zookeeper.PortAssignment;
 import org.apache.zookeeper.ZooKeeper;
 import org.apache.zookeeper.client.ZKClientConfig;
@@ -57,8 +56,6 @@ public class X509AuthFailureTest {
         System.setProperty(clientX509Util.getSslKeystorePasswdProperty(), "testpass");
         System.setProperty(clientX509Util.getSslTruststoreLocationProperty(), testDataPath + "/ssl/testTrustStore.jks");
         System.setProperty(clientX509Util.getSslTruststorePasswdProperty(), "testpass");
-        //System.setProperty("zookeeper.authProvider.x509", "org.apache.zookeeper.server.auth.SFDCX509AuthenticationProvider");
-        //System.setProperty(clientX509Util.getSslProviderProperty(), DynamicKeyStoreProvider.PKIKS_PROVIDER_NAME);
     }
 
     @After
@@ -79,13 +76,16 @@ public class X509AuthFailureTest {
      */
     @Test
     public void testSecureStandaloneServerAuthZFailure() throws Exception {
+
+        final Integer CLIENT_PORT = PortAssignment.unique();
+        final Integer SECURE_CLIENT_PORT = PortAssignment.unique();
         System.setProperty("zookeeper.ssl.allowedRoles", "testauthz");
-        Integer secureClientPort = PortAssignment.unique();
-        MainThread mt = new MainThread(QuorumPeerTestBase.MainThread.UNSET_MYID, "", secureClientPort, false);
+
+        ZooKeeperServerMainTest.MainThread mt = new ZooKeeperServerMainTest.MainThread(CLIENT_PORT, SECURE_CLIENT_PORT, true, null);
         mt.start();
 
         try {
-            ZooKeeper zk = createZKClnt("127.0.0.1:" + secureClientPort);
+            ZooKeeper zk = createZKClnt("127.0.0.1:" + SECURE_CLIENT_PORT);
         } catch (Exception e){
             ServerStats serverStats = mt.getSecureCnxnFactory().getZooKeeperServer().serverStats();
             assertTrue(serverStats.getAuthFailedCount()>=1);
@@ -101,142 +101,4 @@ public class X509AuthFailureTest {
         return zk;
     }
 
-    public static class MainThread extends Thread {
-        final TestZKServerMain main;
-        final File confFile;
-        final File tmpDir;
-
-        public static final int UNSET_STATIC_CLIENTPORT = -1;
-        // standalone mode doens't need myid
-        public static final int UNSET_MYID = -1;
-
-        File baseDir;
-        private int myid;
-        private int clientPort;
-        private Map<String, String> otherConfigs;
-
-        public MainThread(int myid, String quorumCfgSection) throws IOException {
-            this(myid, quorumCfgSection, true);
-        }
-
-        public MainThread(int myid, String quorumCfgSection, Integer secureClientPort, boolean writeDynamicConfigFile)
-                throws  IOException {
-            this(myid, UNSET_STATIC_CLIENTPORT, JettyAdminServer.DEFAULT_PORT, secureClientPort,
-                    quorumCfgSection, null, writeDynamicConfigFile, null);
-        }
-
-        public MainThread(int myid, String quorumCfgSection, boolean writeDynamicConfigFile)
-                throws IOException {
-            this(myid, UNSET_STATIC_CLIENTPORT, quorumCfgSection, writeDynamicConfigFile);
-        }
-
-        public MainThread(int myid, int clientPort, String quorumCfgSection, boolean writeDynamicConfigFile)
-                throws IOException {
-            this(myid, clientPort, JettyAdminServer.DEFAULT_PORT, quorumCfgSection, null, writeDynamicConfigFile);
-        }
-
-        public MainThread(int myid, int clientPort, String quorumCfgSection, boolean writeDynamicConfigFile,
-                          String version) throws IOException {
-            this(myid, clientPort, JettyAdminServer.DEFAULT_PORT, quorumCfgSection, null,
-                    writeDynamicConfigFile, version);
-        }
-
-        public MainThread(int myid, int clientPort, String quorumCfgSection, String configs)
-                throws IOException {
-            this(myid, clientPort, JettyAdminServer.DEFAULT_PORT, quorumCfgSection, configs, true);
-        }
-
-        public MainThread(int myid, int clientPort, int adminServerPort, String quorumCfgSection,
-                          String configs)  throws IOException {
-            this(myid, clientPort, adminServerPort, quorumCfgSection, configs, true);
-        }
-
-        public MainThread(int myid, int clientPort, int adminServerPort, String quorumCfgSection,
-                          String configs, boolean writeDynamicConfigFile)
-                throws IOException {
-            this(myid, clientPort, adminServerPort, quorumCfgSection, configs, writeDynamicConfigFile, null);
-        }
-
-        public MainThread(int myid, int clientPort, int adminServerPort, String quorumCfgSection,
-                          String configs, boolean writeDynamicConfigFile, String version) throws IOException {
-            this(myid, clientPort, adminServerPort, null, quorumCfgSection, configs, writeDynamicConfigFile, version);
-        }
-
-        public MainThread(int myid, int clientPort, int adminServerPort, Integer secureClientPort,
-                          String quorumCfgSection, String configs, boolean writeDynamicConfigFile, String version)
-                throws IOException {
-            tmpDir = ClientBase.createTmpDir();
-            LOG.info("id = " + myid + " tmpDir = " + tmpDir + " clientPort = "
-                    + clientPort + " adminServerPort = " + adminServerPort);
-
-            File dataDir = new File(tmpDir, "data");
-            if (!dataDir.mkdir()) {
-                throw new IOException("Unable to mkdir " + dataDir);
-            }
-
-            confFile = new File(tmpDir, "zoo.cfg");
-
-            FileWriter fwriter = new FileWriter(confFile);
-            fwriter.write("tickTime=4000\n");
-            fwriter.write("initLimit=10\n");
-            fwriter.write("syncLimit=5\n");
-            if(configs != null){
-                fwriter.write(configs);
-            }
-
-            // Convert windows path to UNIX to avoid problems with "\"
-            String dir = PathUtils.normalizeFileSystemPath(dataDir.toString());
-
-            fwriter.write("dataDir=" + dir + "\n");
-            fwriter.write("admin.serverPort=" + adminServerPort + "\n");
-
-            // For backward compatibility test, some tests create dynamic configuration
-            // without setting client port.
-            // This could happen both in static file or dynamic file.
-            if (clientPort != UNSET_STATIC_CLIENTPORT) {
-                fwriter.write("clientPort=" + clientPort + "\n");
-            }
-
-            if (secureClientPort != null) {
-                fwriter.write("secureClientPort=" + secureClientPort + "\n");
-            }
-
-            fwriter.write(quorumCfgSection);
-
-            fwriter.flush();
-            fwriter.close();
-
-            File myidFile = new File(dataDir, "myid");
-            fwriter = new FileWriter(myidFile);
-            fwriter.write(Integer.toString(myid));
-            fwriter.flush();
-            fwriter.close();
-            main = new TestZKServerMain();
-        }
-
-        public void run() {
-            String args[] = new String[1];
-            args[0] = confFile.toString();
-            try {
-                main.initializeAndRun(args);
-            } catch (Exception e) {
-                // test will still fail even though we just log/ignore
-                LOG.error("unexpected exception in run", e);
-            }
-        }
-
-        public void shutdown() {
-            main.shutdown();
-        }
-
-        public ServerCnxnFactory getSecureCnxnFactory(){
-            return main.getSecureCnxnFactory();
-        }
-    }
-
-    public static  class TestZKServerMain extends ZooKeeperServerMain {
-        public void shutdown() {
-            super.shutdown();
-        }
-    }
 }
